+++ changelog.md
@@ -26,8 +26,6 @@ compromise uses semver, and pushes to npm frequently
 - allow ellipses at start of term _"....so"_ in `@hasEllipses`
 - fix matches with optional-end `foo?$` match syntax
 - add typescript types for subsets
-- fix over-active emoji regex
-- tag 'cookin', 'hootin' as `Gerund`
 
 ##### 12.2.1
 
@@ -37,6 +35,9 @@ compromise uses semver, and pushes to npm frequently
 - more tests
 - fix shared period with acronym at end of sentence
 - fix some mis-classification of contraction
+- fix over-active emoji regex
+- tag 'cookin', 'hootin' as `Gerund`
+- support unicode single-quote symbols in contractions
 
 ### v11